extract_utils: add the ability to download from remote with aria2
This will attempt to download the $SRC (if it's .zip) from OEM remote
servers.

Just append the following to extract-files.sh:

´´´
On the top after device and vendor:
REMOTE=http://oem.downloads.com/device

On the bottom, before extraction call:
download "$SRC"
```

Then run:

./extract-files.sh oem_version_device-signed.zip

It will pull down $REMOTE/$SRC from that server and follow up with the
extraction.

This is OPTIONAL and up to the maintainer.

Change-Id: Ic742b3b8b523a4ceb74c2869f23104a425662e2c
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>