extract_utils: clean up the vendor folder properly
The previous change fixed the VENDOR_STATE check but further testing
showed that the VENDOR_STATE check was incorrect for "false" conditions
since it was skipping the copy of the "old blobs" to /tmp and breaking
hash checking on those cases.

This new behaviour uses "cp -r" instead of "mv" when vendor cleaning is
not desired (in most cases) to back up the old blobs to /tmp.

The default value for vendor and radio cleaning is now false, the device
itself should opt-in into that by setting the $5 argument to "true"
on setup_vendor.

Change-Id: I3a76c38cdff69403ea7e56faea67971e8b81e782
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>